// src/app/Models/Ventas.model.ts
export interface Ventas {
  id: number;
  productoId: number;
  productoNombre: string;
  empleadoId: number;
  empleadoNombre: string;
  cantidad: number;
  total: number;
}
// {
//   "id": 0,
//   "productoId": 0,
//   "productoNombre": "string",
//   "empleadoId": 0,
//   "empleadoNombre": "string",
//   "cantidad": 0,
//   "total": 0,
//   "fechaVenta": "2025-10-28T17:52:48.981Z"
// }